requirements.txt
psycopg2-binary
requests
beautifulsoup4
lxml
selenium
playwright
pandas
